+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.7
+
+- Fix create/create_document to support Word (.docx) template automations
+
+## 1.0.6
+
+- Fix create/create_document to support PDF template automations
+
 ## 1.0.2
 
 Added data source "Zapier" in addition to "API".